Tweaked `metalsmith-run` custom module and the `--dist` flag behaviour
The `metalsmith-run` custom module has been simplified to accept a `condition` function that returns a truthy / falsey value in order to determine whether or not to run the `plugin` function. This way, there is no need for any conditional logic to be coded into the `metalsmith-run` module itself.

After that, the `--dist` flag behaviour was tested, checked and tweaked to ensure that none of the serve functions are executed.