Tweaked boilerplate's build function, and added a setup function
1. Rather than having to open the command-line in order to `cd` into
the folder just to run the `node build` command, this is now replaced
with a .bat file. Now all we need to do is double-click on the
`build.bat` file in order to build the site.

2. Rather than having to manually change all `Website Name` as well as
updating the virtual folder name, this is now all done through a guided
setup flow that is triggered by double-clicking on the `setup.bat` file.

3. To really ensure that there are no projects that have a `Website
Name` left as it is, the build command will only run once we've
completed the setup process.