Added port checking before running Metalsmith
To facilitate development of multiple Metalsmith sites at once, the
tool will now check for the availability of ports before proceeding
to run the build. In the event of a conflict, the tool will attempt to
use alternative ports (by simply adding 1 to the port), up to a total
limit of 5 tries, after which the tool stops trying for an alternative
port and just proceeds with the conflicting port.