Added some simple documentation for map-list variables
As map-list variables aren't as straightforward to use in SCSS (needs
to be accessed through `map-get`), there are convenience functions that
are created to be used with these variables... And the only way to know
that is to look at the `_functions.scss` partial file, which is easy to
miss.

The documentation has been added to make it very obvious that there are
already existing convenience functions to make it easier to use these
variables.