Tweaked media query variables
Removed the @each loop that is used to generate the media query
variables. This is to simplify the code and make it more straightforward
for other developers to grasp and modify.

This also has the added benefit of removing the `bp` function, as the
media queries are no longer in a map list.

Feeling good about this particular iteration.