Refactored `metalsmith-default-meta` custom module into `metalsmith-meta-transformers`
Three primary gains from this:

1. The metadata can now be updated on the fly. In the previous
iteration, the build process had to be restarted in order to get the
latest metadata to be reflected in the files.

2. Rather than allowing both strings and functions to be set as the
metadata, it is now all standardised into functions only, and the
function does not need to return the final metadata to be applied onto
the file. This means that the function is now expected to apply all
desired metadata by directly modifying the file.

3. Metadata configurations are now set in an Array, so that the sequence
can be controlled. This allows transformations to be chained in a
specific order.

As it is no longer a straightforward action of setting default metadata,
this custom module has been renamed accordingly.