Tweaked modifier class naming convention
BEM modifiers follow the `.block--modifier` convention, whereas I have
simplified it into `.--modifier`. To make that simplification work, the
first dash has to be escaped, meaning in the SCSS file, it needs to be
written as `.\--modifier`.

The problem is that the escape doesn't work for the print view in IE11.

To work around that, I now place the double dash at the end of the
modifier class name, meaning it should now be `.modifier--`.