Added `metalsmith-data-loader` into the build
The YAML front matter is quite handy for the build, because it allows
lots of templating potential through Nunjucks. Seeing as we may end up
working with lots of data, it would be highly beneficial to have syntax
highlighting for it.

`metalsmith-data-loader` allows us to separate the front matter into
dedicated YAML files, thus allowing us to leverage syntax highlighting.

The main downside of this approach is that saving the YAML file alone
will not result in the changes being reflected in the build's live
reloading feature; after saving the YAML file, we will need to save the
page file too, only then would the latest YAML front matter be loaded
into the page and reflected in the live reloading.